// generated/inputTypeSchemas/AddressUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { OrderUpdateManyWithoutAddressNestedInputSchema } from './OrderUpdateManyWithoutAddressNestedInputSchema';
import { UserUpdateOneRequiredWithoutAddressesNestedInputSchema } from './UserUpdateOneRequiredWithoutAddressesNestedInputSchema';
import { AffiliateUpdateManyWithoutAddressNestedInputSchema } from './AffiliateUpdateManyWithoutAddressNestedInputSchema';

export const AddressUpdateInputSchema: z.ZodType<Prisma.AddressUpdateInput> = z.object({
  id: z.union([ z.string().uuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  phone: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  city: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  district: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  ward: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  street: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  orders: z.lazy(() => OrderUpdateManyWithoutAddressNestedInputSchema).optional(),
  user: z.lazy(() => UserUpdateOneRequiredWithoutAddressesNestedInputSchema).optional(),
  Affiliate: z.lazy(() => AffiliateUpdateManyWithoutAddressNestedInputSchema).optional()
}).strict();

export default AddressUpdateInputSchema;